refactor: move asyncPipe to src/utils and establish subpath export architecture
Major refactoring to prepare for long-term framework vision:

**Package Structure Changes:**
- Moved asyncPipe from lib/ to src/utils/ (lib/ now for CLI internals only)
- Created src/utils/index.js and index.d.ts with TypeScript definitions
- Removed root "." export - framework now uses explicit subpath imports only

**Export Architecture:**
- Added "./package.json" export for version checking
- Added "./utils" subpath export with TypeScript types
- "./server" export already in place with TypeScript types
- No kitchen sink root export - all imports are explicit and tree-shakeable

**Documentation:**
- Added vision.md documenting long-term framework architecture
- Added .plan.md for development roadmap
- Added Utils Library section to README with asyncPipe documentation
- Updated ToC with new sections
- Listed upcoming modules (observables, store, sagas, components, billing, wallet)

**Import Updates:**
- Updated all internal imports to use new src/utils/ path
- Updated examples to use 'aidd/utils' subpath import

All tests passing (87 tests). Ready for future module additions.
