+++ .plan.md
@@ -1,0 +1,44 @@
+# AIDD Development Plan
+
+See [vision.md](./vision.md) for the long-term vision and architecture.
+
+## Current Sprint
+
+### Refactoring for Scale
+- [x] Move asyncPipe from lib/ to src/utils/
+- [x] Set up proper subpath exports (no root export)
+- [x] Add TypeScript definitions for utils
+- [ ] Update README with utils documentation
+- [ ] Commit and push refactoring changes
+
+### Documentation
+- [x] Update position statement from "library" to "framework"
+- [x] Add server framework to README
+- [x] Add CLI help text for server framework
+- [ ] Document utils exports
+- [ ] Add architecture documentation
+
+## Backlog
+
+### Phase 1: Core Infrastructure
+- [ ] Observables API with pipeable operators
+- [ ] Store provider (Supabase auto-sync)
+- [ ] ECS architecture research (adobe/data)
+- [ ] Saga builder foundation
+
+### Phase 2: Developer Experience
+- [ ] DevTools implementation
+- [ ] Time-travel debugging
+- [ ] State inspection tools
+
+### Phase 3: UI & Components
+- [ ] Component library foundation
+- [ ] Authentication components
+- [ ] User preference management
+- [ ] GDPR compliance tools
+
+### Phase 4: Monetization & Web3
+- [ ] Billing system integration
+- [ ] Subscription management
+- [ ] Wallet integration
+- [ ] Payment abstractions
