+++ lib/release-helpers.js
@@ -1,4 +1,4 @@
-import { asyncPipe } from "./asyncPipe.js";
+import { asyncPipe } from "../src/utils/asyncPipe.js";
 import { exec } from "child_process";
 import { promisify } from "util";
