+++ package.json
@@ -20,7 +20,11 @@
     "toc": "doctoc README.md"
   },
   "exports": {
-    ".": "./index.js",
+    "./package.json": "./package.json",
+    "./utils": {
+      "types": "./src/utils/index.d.ts",
+      "default": "./src/utils/index.js"
+    },
     "./server": {
       "types": "./src/server/index.d.ts",
       "default": "./src/server/index.js"
