+++ src/utils/index.js
@@ -1,0 +1,5 @@
+/**
+ * Utility functions for functional composition
+ */
+
+export { asyncPipe } from "./asyncPipe.js";
